+++ DbViewer.TestApi/Impl/SampleObjectsSearcher.cs
@@ -2,7 +2,6 @@
 
 using SkbKontur.DbViewer.Connector;
 using SkbKontur.DbViewer.DataTypes;
-using SkbKontur.DbViewer.TestApi.Impl.Classes;
 
 #pragma warning disable 1998
 
@@ -13,28 +12,27 @@ namespace SkbKontur.DbViewer.TestApi.Impl
     {
         public async Task<object[]> Search(Condition[] filters, Sort[] sorts, int from, int count)
         {
-            // ReSharper disable once CoVariantArrayConversion
-            return SampleDataBase.Instance.Find(filters, sorts, from, count);
+            return SampleDataBase.Get<T>().Find(filters, sorts, from, count);
         }
 
         public async Task<int?> Count(Condition[] filters, int? limit)
         {
-            return SampleDataBase.Instance.Count(filters, limit);
+            return SampleDataBase.Get<T>().Count(filters, limit);
         }
 
         public async Task<object> Read(Condition[] filters)
         {
-            return SampleDataBase.Instance.Read(filters);
+            return SampleDataBase.Get<T>().Read(filters);
         }
 
         public async Task Delete(Condition[] filters)
         {
-            SampleDataBase.Instance.Delete(filters);
+            SampleDataBase.Get<T>().Delete(filters);
         }
 
         public async Task Write(object @object)
         {
-            SampleDataBase.Instance.Write((TestClass)@object);
+            SampleDataBase.Get<T>().Write(@object);
         }
     }
 }